fix pylint in _icom
